Use a random control token by default
Closes #25
Fixes #25

I'm still not convinced this is really necessary, but I don't have a
good argument for not doing it either. If this ends up being a PITA for
people that happen to use the control token as part of other tests, then
we'll revisit plopping it somewhere.

I suppose they could always do their own thing in the `Guardfile`...